+++ src/main/java/org/openrewrite/java/spring/boot3/AddRouteTrailingSlash.java
