+++ src/main/java/org/openrewrite/java/spring/boot3/MaintainTrailingSlashURLMappings.java
@@ -47,7 +47,7 @@ public class MaintainTrailingSlashURLMappings extends Recipe {
             return before;
         }
 
-        doNext(new AddMappingPathsWithTrailingSlash());
+        doNext(new AddRouteTrailingSlash());
         return before;
     }
